+++ spec/unit/web/test_magic_assets.rb
@@ -1,5 +1,6 @@
 require 'spec_helper'
 require 'startback/web/magic_assets'
+require 'startback/web/magic_assets/ng_html_transformer'
 
 module Startback
   module Web
@@ -21,7 +22,7 @@ module Startback
 
         it 'delegates a [] call to sprockets' do
           result = app['index.js']
-          expect(result.to_s).to match(/function test/)        
+          expect(result.to_s).to match(/function test/)
         end
 
         it 'returns a 404 on unknown' do
@@ -53,6 +54,26 @@ module Startback
           get "/my-assets/index.js"
           expect(last_response.status).to eql(200)
           expect(last_response.body).to match(/function test/)
+        end
+      end
+
+      context 'when registering the NgHtmlTransformer' do
+        let(:app){
+          plugin = MagicAssets::NgHtmlTransformer.new({
+            ng_module: "test.templates"
+          })
+          MagicAssets.new({
+            folder: Path.dir/"fixtures/assets",
+            plugins: [plugin]
+          })
+        }
+
+        it 'works as expected' do
+          get "/index.js"
+          expect(last_response.status).to eql(200)
+          expect(last_response.body).to match(/function test/)
+          expect(last_response.body).to match(/Hello \{\{who\}\}/)
+          expect(last_response.body).to match(/angular.module\("test.templates"\)/)
         end
       end